Re-implement delimiterChars in onInput callback
This re-implements the delimiterChars option as an array of literal characters. When the entered
query increases in length the last character is checked against this list. This replaces the use
of keyboard events to find out which key was pressed because key mapping does not always occur
when entry is made by a virtual or soft keyboard.